commands: rename rooted path conversion function
Since the "git lfs checkout" command was introduced in PR #527, it has
accepted zero or more command-line arguments which are expected to be
file path patterns.  If any such arguments are supplied, they are used
to filter the set of Git LFS files that will be processed by the command.

When the command is run in a subdirectory within a repository's working
tree, we expect that the user will provide path patterns which are
expressed relative to their current working directory.  However, as
noted in commit 760c7d75e1cb4ea95f2a80f0ecf6d45620d56624 of PR #527,
we need to convert these path patterns so they are relative to the
root of the repository before we can use them to filter the file list
returned by Git.  (This list is either generated by a "git ls-files"
command, if the installed version of Git is 2.42.0 or higher, or by
a "git ls-tree" command.)

In PR #1771 we refactored the logic we use to perform these relative
path conversions and created the rootedPath() helper function in our
"commands" package.  This function iterated over the set of
command-line "pathspec" arguments and converted each using the
Convert() method of the currentToRepoPathConverter structure in our
"lfs" package.

Later, in commit 56abb7122a7aa62408d0a14e23907d88b431abdf of PR #4556,
we revised the rootedPaths() function to instead call the Convert()
method of the new currentToRepoPatternConverter structure in our "lfs"
package.  This structure's conversion method first calls the Convert()
method of a currentToRepoPathConverter structure, but then checks
whether the result references a directory or not, and if so, adjusts
the returned value to conform to Git's pattern-matching rules, as
defined in the gitignore(5) manual page.

While this treatment of path pattern arguments is correct for most use
cases, it is not valid when the "git lfs checkout" command's --to option
is specified by the user, but the rootedPaths() function is nevertheless
used to process the command's final file path argument in this case.

In a subsequent commit in this PR we will address this problem, but
first we rename the rootedPaths() function to rootedPathPatterns()
and make the same change to the equivalent names of the local variables,
so as to clearly identify the purpose of the function and its expected
input and output.