t/t-checkout.sh: skip link checks unless supported
In commit 5c11ffce9a4f095ff356bc781e2a031abb46c1a8 of our "main"
development branch and commit 8b4aede153ff6555a7ad96deb2d57c333fdb446d
of our "release-3.7" branch we revised the SmudgeToFile() method of
the GitFilter structure in our "lfs" package so it removes any existing
directory entry at the path where it intends to create a file before
attempting to create that file.  Among other advantages, this change
ensures that we will not write through a symbolic link which exists
in place of the file we intend to create, and so forms part of our
remediation of the vulnerabilities reported as CVE-2025-26625.

When we made this change we updated several of the tests in our
t/t-checkout.sh and t/t-pull.sh shell test scripts, including the
"checkout: conflicts" test, which validates the behaviour of the
"git lfs checkout" command's --to option.  In particular, we revised
this test so that it now confirms that any pre-existing symbolic links
at the file path specified with the --to option are removed before a
new file is written at the given path.

However, on Windows these checks are only useful if true symbolic link
support is enabled.  In an earlier pair of commits, commit
7a86d13d64ae69ea760cc96f0cfd01acee2fc46a of our "main" branch and commit
1b483dbf97fc6de1d5b2046c1af1e336b66a3fc5 of our "release-3.7" branch,
we expanded the "checkout: conflicts" test and introduced a check which
confirms that the "git lfs checkout" command will traverse a symbolic
link in the path provided with the --to option so long as it appears
in place of a directory in the path and links to an existing directory.

On Windows, though, we only perform this check when the
has_native_symlinks() test helper function returns a successful (i.e.,
zero) exit code.  When we further expanded the "checkout: conflicts"
test in the later commits 5c11ffce9a4f095ff356bc781e2a031abb46c1a8
and 8b4aede153ff6555a7ad96deb2d57c333fdb446d, though, we overlooked
this conditional block.  As a result, our checks that symbolic links
are removed when they appear in place of the file identified by the
--to option always run on Windows, even when symbolic links are merely
simulated by the MSYS2 environment using deep copies of their target
locations.

We therefore now adjust the "checkout: conflicts" test so that all
of its checks of the "git lfs checkout" command's handling of symbolic
links in the path specified by the --to option only run on Windows
when such links are actually supported by the operating system and
MSYS2 environment.