commands,t: convert merge conflict paths correctly
The --to option of the "git lfs checkout" command requires a file
path argument which specifies where the command should write the
contents of the object associated with a Git LFS pointer file.  The
command also requires another file path argument which identifies
the pointer within the repository whose object contents should be
retrieved, along with a second option that specifies which stage of
a merge conflict should be referenced when reading the pointer's data.
This second option must be one of --theirs, --ours, or --base.

The final file path argument, which indicates the pointer file within
the repository whose object contents are to be retrieved, may be
provided by the user as a relative path from their current working
directory.  This path is then converted into a path relative to the
root of the repository by the rootedPathPatterns() function, which calls
the Convert() method of the currentToRepoPatternConverter structure in
our "lfs" package.  However, because only a single file path is expected
and not a file path pattern, the Convert() method of the
currentToRepoPathConverter structure should be used instead.

The Convert() method of the currentToRepoPatternConverter structure
is intended to process file path patterns and to make sure they adhere
to Git's pattern-matching rules, as defined in the gitignore(5) manual
page.  Paths which resolve to directories have "/" appended by the
currentToRepoPatternConverter structure's Convert() method so that they
will match the directory and all of its contents.  Any leading "./" path
component is then removed, and if this results in an empty pattern,
then "**" is returned so that the current directory and all of its
contents will be matched by the pattern.

As noted above, though, when the --to option is used along with one
of the --theirs, --ours, or --base options, neither of the two required
file path arguments should be a file path pattern, and in particular
the second file path argument should uniquely identify the Git LFS file
that is the source of a merge conflict.

Since this second file path may be specified as a path relative to
the current working directory, the "git lfs checkout" command needs
to convert it to be relative to the root of the repository, but without
treating it as a file path pattern.  We therefore revise the command
so that when the --to option is given, the file path argument which
identifies a Git LFS file in the repository is processed using the
Convert() method of the currentToRepoPathConverter structure rather
than the Convert() method of the currentToRepoPatternConverter
structure.

We then add a check to the "checkout: conflicts" test in our
t/t-checkout.sh test script which verifies that our changes have the
intended effect.  To do this, we run the "git lfs checkout" command
with the --to option, and for the file path parameter of the --ours
option we provide a simple "." path.  Previously, this would have
been converted into "**", but now we should see just the original "."
character in the error message returned from the "git rev-parse"
command.  (The "git lfs checkout" command uses this Git command to
determine the SHA of the merge stage containing the file path by
prepending the stage's numeric identifier to the path.  This presumes,
of course, that the user has encountered a merge conflict on a Git LFS
file and is now using the "git lfs checkout" command's --to option to
help resolve the conflict.)