t: check repo names in assertions and test helpers
In a prior commit in this PR we revised a number of the test helper
functions and assertion functions in our shell test suite so they now
verify that they are passed a valid Git LFS object ID in their "oid"
parameters.

Several of these functions also accept a "reponame" parameter, and
in most cases the functions effectively confirm that parameter is
not empty or invalid simply through their normal action.  For instance,
the refute_remote_object() function expects to temporarily change the
working directory to a location whose path contains the "reponame"
parameter, and so the function will fail if that parameter is incorrect.

Likewise, the "positive" assertion functions assert_server_object()
and assert_remote_object() can only succeed when the given repository
name actually corresponds to an appropriate storage location.

However, at present the refute_server_object() and delete_server_object()
functions do not fail if an invalid repository name is provided to them.

The delete_server_object() helper function makes a "DELETE" HTTP request
to the test server, whose lfsDeleteHandler() function always returns a
200 status code even if the request URL does not correspond to any Git LFS
object stored in the server's memory.  We therefore adjust this function
so that it checks for a Git repository directory with the same base name
as the "reponame" parameter (plus the ".git" extension) in the location
used by the test server.  If such a directory does not exist, the
helper function will now fail.

The refute_server_object() assertion function, meanwhile, makes a "GET"
HTTP request to the test server, and fails if the server returns a JSON
response indicating that the given Git LFS object is available for
download.  If an empty or invalid repository name parameter is provided,
the server will not find any Git LFS object corresponding to the
request URL constructed with the incorrect repository name, and so the
assertion will succeed for the wrong reason.  Hence we also adjust this
function to check for a Git repository directory in the same manner as
the delete_server_object() function.

We then need to update the three tests in our t/t-credentials-protect.sh
test script so that they create Git repository directories with the
same repository names that they use in the "lfs.url" configuration
option, or else the refute_server_object() assertions in these tests
will now fail because those Git directories do not exist.