t/t-no-remote.sh: fix tar archive checks
In commits 2e924e04a2a928567f83f1a69d2a42722b02ff49 and
e09b7b07fee4d85856919a581ed928f4a2ee9aa8 of PR #5357 we created the
t/t-no-remote.sh test script and added its two tests.  Both of these
tests end by generating a "tar" archive file with the "git archive"
command and then trying to check its contents.

At present, these checks use the grep(1) command to look for strings
in the "tar" archive file.  However, the file is technically not a text
file, so using the "grep" command to search for text strings in it
is not ideal.

Moreover, the final check in the "fallback ignored when remote present"
test attempts to verify that a given file is not included in the contents
of the archive file by using the "grep" command with the -v option.

However, the -v option of the "grep" command does not cause the command
to fail (i.e., exit with a non-zero value) if a line is found in its
input which matches the provided regular expression.  Rather, the -v
option causes the "grep" command to filter out any lines from its input
which match the expression, and then the exit status is determined in
the usual manner, so that the command only returns a non-zero value if
no other lines were seen in the input.

Since the archive file always contains many strings and newline
characters, when the "grep" command is used with the -v option it
finds content other than the provided file name and so always succeeds.
Even if the file name did unexpectedly appear in the contents of the
archive file, it would simply be filtered by the "grep" command because
of the use of the -v option, and then the existence of all the other
content in the archive file would allow the "grep" command to succeed.

To resolve this problem and also avoid searching a binary file format
as if it was a text file, we replace the use of the "grep" command
in both tests in the t/t-no-remote.sh script.  Instead of attempting
to search the contents of the archive file, we use the tar(1) command
itself to list the files in the archive which match a given filename,
and then count those lines with the wc(1) command and check that the
count matches the value we expect, either one or zero.