t/t-track.sh: fix incorrect grep -v usage
In commit 226f5bb092125cab760759863219d125c6432592 of PR #3712 we
updated the "git lfs track" command to respect the "core.autocrlf"
Git configuration option setting of "input".

In particular, we revised the command so that when it is not possible to
determine the line-ending format of a ".gitattributes" file (either
because the file does not exist or because it does not contain any
CR or LF characters), the "git lfs track" command will use the default
line-ending format of the current operating system when rewriting the
".gitattributes" file.

(In PR #3712 we did not, however, update the "git lfs untrack" command
to respect the "core.autocrlf" setting in the same manner.  Further, we
did not address the problem that the "git lfs untrack" command always
rewrites ".gitattributes" files using only LF line-ending characters,
even on Windows or when the file already contains line endings with
pairs of CR and LF characters.  In order to keep the changes in this
PR constrained to our test suite, though, we will defer correction of
these issues to a future PR.)

In addition to the changes made to the "git lfs track" command in
commit 226f5bb092125cab760759863219d125c6432592 of PR #3712, we also
updated the "track with autocrlf=input" test in our t/t-track.sh test
script so that it attempts to check that the correct EOL (End-Of-Line)
characters are present in the ".gitattributes" file after the
"git lfs untrack" command is run.

To check the EOL characters in the modified ".gitattributes" file,
the test now uses the cat(1) command's -e option to output the file's
contents with special characters (other than the LF character) displayed
in visible format and a dollar sign character at the end of each line,
and then runs a grep(1) command on that output to try to verify that
it contains the expected EOL characters for a given OS platform.

On Windows, we use the "grep" command to confirm that the ".gitattributes"
file contains a trailing CR character before the final LF character on
at least one line, and on non-Windows systems, we use the "grep" command
with its -v option to try to confirm that no CR characters are present
in any of the file's lines.

However, the -v option of the "grep" command does not cause the command
to fail (i.e., exit with a non-zero value) if a line is found in its
input which matches the provided regular expression.  Rather, the -v
option causes the "grep" command to filter out any lines from its input
which match the expression, and then the exit status is determined in
the usual manner, so that the command only returns a non-zero value if
no other lines were seen in the input.

Since the test creates two lines in the ".gitattributes" file, if the
"git lfs track" command rewrote the file such that one line had only a
LF line-ending character while the other had both CR and LF line-ending
characters, then when the "grep" command was used with the -v option,
the line with the CR character would be filtered and the existence of
the other line would still allow the "grep" command to succeed.

Also, given such an input file with mismatched EOL characters on
different lines, the "grep" command we run on Windows would still succeed
even though one of the lines lacked a CR character before the final LF
character.

To resolve both of these issues, we revise the test so that it uses
the -c option of the "grep" command in both cases, without the -v option,
and then checks that the line counts returned by the "grep" commands
match the expected values.  On Windows, this means we expect two lines,
both with CR and LF line-ending characters, while on non-Windows systems
we expect zero lines to contain any CR characters.

As described in a prior commit in this PR, we already use this idiom
throughout many of our other test scripts, in part because it has several
advantages over other possible techniques for ensuring that a file
contains no lines (or an exact number of lines) matching a certain
pattern.  In particular, if the "grep" command were to fail with an
error, it would not output an integer count value, which would cause
the test to fail.

We also take the opportunity to adjust the indentation of the code in
both cases so that they match our usual style of two space characters
per level of indentation.