t: always use grep -c when no matches expected
In several prior commits in this PR we addressed the instances in our
shell test suite where we used the grep(1) command with the invalid
assumption that the command would fail if one of its input lines matched
the pattern provided with the -v option.

We corrected those use cases by revising them to use the -c option of
the "grep" command rather than the -v option, and then checking that the
number of lines the command finds which match the given pattern is zero.

As described in our prior commits in this PR, we already use this idiom
throughout many of our other test scripts, in part because it has several
advantages over other possible techniques for ensuring that the input to
the "grep" command contains no lines matching a certain pattern.  In
particular, if the "grep" command were to fail with an error, it would
not output an integer count value, which would cause the test to fail.

One other technique we use for this same purpose is to check that the
output of the "grep" command is an empty string, either using the -z
shell test or, in the case of the assert_dir_perms() function in our
t/t-umask.sh test script, with the shell string comparison operator.

Under normal circumstances, the checks using the -z or string comparison
shell test operators work as expected.  However, if a "grep" command
was to fail with an error and return a status code of 2, this would not
trigger a failure of the enclosing test, for a combination of reasons.

Although the "errexit" shell attribute is enabled in all of our tests
by the "set -e" command, because we run the "grep" command in a subshell
within the shell "[" builtin, if the command returns a non-zero status
code that will not cause the test's shell to exit immediately.  The
command would also likely print an error message to its standard error
file descriptor but would not write anything to its standard output, so
our checks using the -z or string comparison shell test operators would
still succeed.

To better guard against regressions where we accidentally introduce an
invalid option or other error into the "grep" commands of these checks,
we revise them now to use our preferred idiom.  Specifically, we add
the -c option to all of the "grep" commands in these checks, and then
revise the checks so that they verify that the "grep" commands report
that no lines in their input matched their given patterns.

Note that in four instances we use the -v option of the "grep" command
correctly, because we intend to filter specific lines from the input
and then confirm that no other lines are present in the input.

These four instances include three in the "track blocklisted files by
name" and "track blocklisted files with glob" tests of the t/t-track.sh
test script where we ignore lines output by "git status --porcelain"
commands when they begin with a pair of question mark characters.

The remaining instance is in the assert_clean_worktree_with_exceptions()
helper function that we added to the t/testhelpers.sh shell library in
commit f3787d66ced28b4244eafc29abb53bab2525f60f of our "main" development
branch and commit df3d0aa61cb9b224fdecf4106e8c0e9d3a52abc6 of our
"release-3.7" branch.  In this function, we ignore specific lines
output by a "git diff-index HEAD" command using a regular expression
supplied by the calling test.

In all four of these instances, we do not need to remove the -v option
of the "grep" command, as we are using it correctly to filter lines
rather than try to demonstrate that certain lines do not exist in
the command's input.